fix: improve error handling for streaming responses
If an error like `StopIteration` is thrown in _do_streaming_response,
the HTTP response does not contain any of the error details and
the grpc status defaults to `2 - Unknown`.

This change makes some of the header/trailer setup more reusable and
passes the trailer data to the response in the error case.